Added tests and code for sorted and summed
The two other methods aside from looping are to sum the list and take
away the result from the sum of the natural list, and to sort the list
and find the number missing.
Summing the list is the best and most efficient way.
Insertion sort code isn't mine. I studied it, I implemented it once or
twice, but for ease I just took it from elsewhere as it wasn't the focus
of this playing around.